Move ErrorCategory and ErrorCode out from the sync package to interop
