+++ packages/cinterop/src/nativeDarwin/kotlin/io/realm/kotlin/internal/interop/ErrorCategory.kt
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package io.realm.kotlin.internal.interop.sync
+package io.realm.kotlin.internal.interop
 
 import realm_wrapper.realm_error_category
