+++ packages/cinterop/src/nativeDarwin/kotlin/io/realm/kotlin/internal/interop/sync/SyncErrorCodeCategory.kt
@@ -16,6 +16,7 @@
 
 package io.realm.kotlin.internal.interop.sync
 
+import io.realm.kotlin.internal.interop.CodeDescription
 import realm_wrapper.realm_sync_error_category
 
 actual enum class SyncErrorCodeCategory(
